gather powered tools recipes in PoweredTools.groovy
