+++ groovy/postInit/mod/MachineRecipes.groovy
@@ -1188,62 +1188,6 @@ ASSEMBLER.recipeBuilder()
     .duration(600)
     .buildAndRegister() 
 
-//Power Units
-ASSEMBLER.recipeBuilder()
-    .inputs(ore('gearSmallSteel') * 2)
-    .inputs(ore('screwSteel'))
-    .inputs(ore('plateSteel') * 2)
-    .inputs(metaitem('electric.motor.lv'))
-    .inputs(metaitem('battery.lead_acid'))
-    .outputs(metaitem('power_unit.lv').withNbt(['MaxCharge': Globals.batteryCapacities['lead_acid']]))
-    .EUt(30)
-    .duration(150)
-    .buildAndRegister()
-
-ASSEMBLER.recipeBuilder()
-    .inputs(ore('gearSmallAluminium') * 2)
-    .inputs(ore('screwAluminium'))
-    .inputs(ore('plateAluminium') * 2)
-    .inputs(metaitem('electric.motor.mv'))
-    .inputs(ore('batteryMv'))
-    .outputs(metaitem('power_unit.mv'))
-    .EUt(120)
-    .duration(150)
-    .buildAndRegister()
-
-ASSEMBLER.recipeBuilder()
-    .inputs(ore('gearSmallStainlessSteel') * 2)
-    .inputs(ore('screwStainlessSteel'))
-    .inputs(ore('plateStainlessSteel') * 2)
-    .inputs(metaitem('electric.motor.hv'))
-    .inputs(ore('batteryHv'))
-    .outputs(metaitem('power_unit.hv'))
-    .EUt(480)
-    .duration(150)
-    .buildAndRegister()
-
-ASSEMBLER.recipeBuilder()
-    .inputs(ore('gearSmallTitanium') * 2)
-    .inputs(ore('screwTitanium'))
-    .inputs(ore('plateTitanium') * 2)
-    .inputs(metaitem('electric.motor.ev'))
-    .inputs(ore('batteryEv'))
-    .outputs(metaitem('power_unit.ev'))
-    .EUt(1920)
-    .duration(150)
-    .buildAndRegister()
-
-ASSEMBLER.recipeBuilder()
-    .inputs(ore('gearSmallTungstenSteel') * 2)
-    .inputs(ore('screwTungstenSteel'))
-    .inputs(ore('plateTungstenSteel') * 2)
-    .inputs(metaitem('electric.motor.iv'))
-    .inputs(ore('batteryIv'))
-    .outputs(metaitem('power_unit.iv'))
-    .EUt(7680)
-    .duration(150)
-    .buildAndRegister()
-
 // AIR DIST. CONTROLLERS
 
 ASSEMBLER.recipeBuilder()
